* Update 3-alpha source fit functions and procedure

+++ NPLib/Calibration/NPSiliconCalibrator.h
@@ -24,7 +24,7 @@ namespace NPL{
 
     public:
       // Use the Zero Extrapolation method to perform fit and return the dead layer thickness
-      double ZeroExtrapolation(TH1* histo, NPL::CalibrationSource* CS, NPL::EnergyLoss* EL, vector<double>& coeff, unsigned int pedestal, unsigned int max_iteration = 10000 , double rmin=-1,double rmax=-1);
+      double ZeroExtrapolation(TH1* histo, NPL::CalibrationSource* CS, NPL::EnergyLoss* EL, vector<double>& coeff, double pedestal, unsigned int max_iteration = 10000 , double rmin=-1,double rmax=-1);
 
       double SimpleCalibration(TH1* histo, NPL::CalibrationSource* CS, NPL::EnergyLoss* EL, vector<double>& coeff, double Assume_Thickness=0,double rmin=-1,double rmax=-1);
